Fix bug 1709540 -- Board frame not always redrawn properly
The bottom and right borders of the chess board are actually
drawn 1 pixel outside of the actual 1st rank and h-file squares
(by design).  But the region-intersects test that controls
which squares get redrawn during a paint event fails to
account for this.  So here, we just unconditionally redraw
these borders so that small artifacts aren't left on the
screen.