Fix En passant handling.
After the move 1. e4, the square e3 should _not_ be marked as
the en passant square in a FEN (or anywhere else).  The only
time an ep. square should be recorded is if a pawn jumps over
it while attacked by an opponent pawn.

This change affects the hash calculation so the binary eco
hashes are updated here as well.