Base guess-the-move suggestions on ECO data.
Precalculate a new set of data that can be loaded when ChessX
starts.  Based on the chessx.eco data, it allows move guesser
to do a simple lookup to see if the ECO data suggests a move
for a given square (ie. the one the mouse is hovering over)
relative to the current board position.

If a move is found it is displayed without delay.  Only if no
move is returned, does move guesser need to invoke the internal
chess engine to make a guess.

Compileeco is used to generate the data file used by guess the
move.  When it finds multiple moves in the ECO data for a given
position-square it uses the chess engine to decide which move
should be added to the chessx.gtm file.  This isn't ideal, and
should be changed to have the decision based on frequency data
instead.