change struct `key_compare` to compile on gcc-14 on fedora 40
Fedora 40 build on gcc-14.x failed when compiling polyglotdatabase.cpp
There is no need to inherit the deprecated std::binary_function.

A simple functor class is OK here (and used only in this file so no
problem) and should work with any environment and compiler.

Same error occurs on fc-38/39 and gcc-13.x