Re #211 New tool window shows centipawn loss graph
There were two gametoolbars. One GameToolBar in
src/gui/gametoolbar.{cpp,h}, and a certain MainWindow::gameToolBar. The
former was a toolbar attached to the notation window showing a material
graph and some clocks shown when "game time" was selected. The latter is
a toolbar at the top of the main window showing "game"-related actions
(such as flipping the board, etc.). The latter was not affected in this
commit.

Limit evaluations to +/- 10.0