Re #211 Select move from game time properly
There was a problem with how gametime move selection worked:

1. Click a spot on game time (which selects a move in the game)
2. Press right arrow a couple times to select a different move
3. Click the same spot on game time

This should rewind the game back to the spot clicked in game time, but
if you didn't move the mouse between steps 1. and 3., the move wasn't
updated.

Another bug is fixed: any time the user clicked a spot in game time, the
move was updated twice: once on mouse "press" and once on mouse
"release".