Re #225 Created dockable window for duplicate positions finder
When preparing an opening repertoine, it is important to note duplicate
positions reached through transposition. When such a transposition is
noted, if there is more than one continuation, the opening book becomes
conflicting and suboptimal. Even if the continuations after the
duplicated position match, the opening book is in danger of conflicting
with itself.

The added Duplicate Positions finder makes it very easy to see at a
glance any such duplicated positions and whether they potentially
conflict or not.

Clicking on a move list for a duplicated position opens that position on
the main board.